Move NetworkMonitor to data/network folder
