+++ app/src/androidTest/java/com/github/geohunt/app/model/database/NetworkMonitorTest.kt
@@ -1,6 +1,7 @@
 package com.github.geohunt.app.model.database
 
 import androidx.compose.ui.test.junit4.createComposeRule
+import com.github.geohunt.app.data.network.NetworkMonitor
 import com.github.geohunt.app.utils.assertFinishes
 import com.github.geohunt.app.utils.assertTimesOut
 import com.google.firebase.database.FirebaseDatabase
